+++ src/axom/quest/tests/quest_sampling_shaper.cpp
@@ -40,6 +40,7 @@ namespace primal = axom::primal;
 namespace quest = axom::quest;
 namespace sidre = axom::sidre;
 namespace slic = axom::slic;
+namespace fs = axom::utilities::filesystem;
 
 namespace
 {
@@ -61,39 +62,6 @@ const std::string proe_tet_fmt_str = R"(
 
 // Set the following to true for verbose output and for saving vis files
 constexpr bool very_verbose_output = false;
-
-/// RAII utility class to write a file at construction time and remove it
-/// once the instance is out of scope
-class ScopedTemporaryFile
-{
-public:
-  ScopedTemporaryFile(const std::string& filename, const std::string& contents)
-    : m_filename(filename)
-  {
-    std::ofstream ofs(m_filename.c_str(), std::ios::out);
-    ofs << contents;
-  }
-
-  ~ScopedTemporaryFile() { EXPECT_EQ(axom::utilities::filesystem::removeFile(m_filename), 0); }
-
-  const std::string& getFileName() const { return m_filename; }
-
-  std::string getFileContents() const
-  {
-    std::stringstream buffer;
-
-    std::ifstream ifs(m_filename.c_str(), std::ios::in);
-    if(ifs.is_open())
-    {
-      buffer << ifs.rdbuf();
-    }
-
-    return buffer.str();
-  }
-
-private:
-  const std::string m_filename;
-};
 
 // Utility function to slice a tetrahedron along a plane
 primal::Polygon<double, 3> slice(const primal::Tetrahedron<double, 3>& tet,
@@ -559,33 +527,6 @@ public:
 
 //-----------------------------------------------------------------------------
 
-TEST(ScopedTemporaryFile, basic)
-{
-  using axom::utilities::filesystem::pathExists;
-
-  const std::string filename = "previously_nonexistent.file";
-  const std::string contents = "file contents!";
-
-  // File does not exist before entering scope
-  EXPECT_FALSE(pathExists(filename));
-
-  // File is created and exists within the scope
-  {
-    ScopedTemporaryFile test_file(filename, contents);
-
-    EXPECT_EQ(test_file.getFileName(), filename);
-    ASSERT_TRUE(pathExists(test_file.getFileName()));
-
-    // check contents
-    EXPECT_EQ(contents, test_file.getFileContents());
-  }
-
-  // File no longer exists outside the scope
-  EXPECT_FALSE(pathExists(filename));
-}
-
-//-----------------------------------------------------------------------------
-
 TEST_F(SamplingShaperTest2D, check_mesh)
 {
   auto& mesh = this->getMesh();
@@ -622,12 +563,12 @@ shapes:
 
   const std::string circle_material = "circleMat";
 
-  ScopedTemporaryFile contour_file(axom::fmt::format("{}.contour", testname), unit_circle_contour);
+  fs::TempFile contour_file(testname, ".contour");
+  contour_file.write(unit_circle_contour);
 
-  ScopedTemporaryFile shape_file(axom::fmt::format("{}.yaml", testname),
-                                 axom::fmt::format(axom::fmt::runtime(shape_template),
-                                                   circle_material,
-                                                   contour_file.getFileName()));
+  fs::TempFile shape_file(testname, ".yaml");
+  shape_file.write(
+    axom::fmt::format(axom::fmt::runtime(shape_template), circle_material, contour_file.getPath()));
 
   if(very_verbose_output)
   {
@@ -635,8 +576,8 @@ shapes:
     SLIC_INFO("Shape file: \n" << shape_file.getFileContents());
   }
 
-  this->validateShapeFile(shape_file.getFileName());
-  this->initializeShaping(shape_file.getFileName());
+  this->validateShapeFile(shape_file.getPath());
+  this->initializeShaping(shape_file.getPath());
   this->runShaping();
 
   // check that the result has a volume fraction field associated with the circle material
@@ -670,15 +611,15 @@ shapes:
 
   const std::string circle_material = "circleMat";
 
-  ScopedTemporaryFile contour_file(axom::fmt::format("{}.contour", testname), unit_circle_contour);
+  fs::TempFile contour_file(testname, ".contour");
+  contour_file.write(unit_circle_contour);
 
-  ScopedTemporaryFile shape_file(axom::fmt::format("{}.yaml", testname),
-                                 axom::fmt::format(axom::fmt::runtime(shape_template),
-                                                   circle_material,
-                                                   contour_file.getFileName()));
+  fs::TempFile shape_file(testname, ".yaml");
+  shape_file.write(
+    axom::fmt::format(axom::fmt::runtime(shape_template), circle_material, contour_file.getPath()));
 
-  this->validateShapeFile(shape_file.getFileName());
-  this->initializeShaping(shape_file.getFileName());
+  this->validateShapeFile(shape_file.getPath());
+  this->initializeShaping(shape_file.getPath());
 
   // check that we can set several projectors in 2D and 3D
   // uses simplest projectors, e.g. identity in 2D and 3D
@@ -722,15 +663,15 @@ shapes:
 
   const std::string circle_material = "circleMat";
 
-  ScopedTemporaryFile contour_file(axom::fmt::format("{}.contour", testname), unit_circle_contour);
+  fs::TempFile contour_file(testname, ".contour");
+  contour_file.write(unit_circle_contour);
 
-  ScopedTemporaryFile shape_file(axom::fmt::format("{}.yaml", testname),
-                                 axom::fmt::format(axom::fmt::runtime(shape_template),
-                                                   circle_material,
-                                                   contour_file.getFileName()));
+  fs::TempFile shape_file(testname, ".yaml");
+  shape_file.write(
+    axom::fmt::format(axom::fmt::runtime(shape_template), circle_material, contour_file.getPath()));
 
-  this->validateShapeFile(shape_file.getFileName());
-  this->initializeShaping(shape_file.getFileName());
+  this->validateShapeFile(shape_file.getPath());
+  this->initializeShaping(shape_file.getPath());
 
   // check that we can set several projectors in 2D and 3D
   // creating an ellipse by scaling input x and y by scale_a and scale_b
@@ -783,11 +724,11 @@ shapes:
   const std::string outer_material = "outer";
   const std::string inner_material = "inner";
 
-  ScopedTemporaryFile contour_file(axom::fmt::format("{}.contour", testname), unit_circle_contour);
+  fs::TempFile contour_file(testname, ".contour");
+  contour_file.write(unit_circle_contour);
 
-  ScopedTemporaryFile shape_file(
-    axom::fmt::format("{}.yaml", testname),
-    axom::fmt::format(axom::fmt::runtime(shape_template), contour_file.getFileName()));
+  fs::TempFile shape_file(testname, ".yaml");
+  shape_file.write(axom::fmt::format(axom::fmt::runtime(shape_template), contour_file.getPath()));
 
   if(very_verbose_output)
   {
@@ -795,9 +736,9 @@ shapes:
     SLIC_INFO("Shape file: \n" << shape_file.getFileContents());
   }
 
-  this->validateShapeFile(shape_file.getFileName());
+  this->validateShapeFile(shape_file.getPath());
 
-  this->initializeShaping(shape_file.getFileName());
+  this->initializeShaping(shape_file.getPath());
   this->runShaping();
 
   // check that the result has a volume fraction field associated with the circle material
@@ -843,16 +784,17 @@ shapes:
       - scale: .5
 )";
 
-  ScopedTemporaryFile contour_file(axom::fmt::format("{}.contour", testname), unit_circle_contour);
+  fs::TempFile contour_file(testname, ".contour");
+  contour_file.write(unit_circle_contour);
 
   // Set background material to 'void' (which is not present elsewhere)
   {
-    ScopedTemporaryFile shape_file(axom::fmt::format("{}.yaml", testname),
-                                   axom::fmt::format(axom::fmt::runtime(shape_template),
-                                                     contour_file.getFileName(),
-                                                     "void",
-                                                     "disk",
-                                                     "hole"));
+    fs::TempFile shape_file(testname, ".yaml");
+    shape_file.write(axom::fmt::format(axom::fmt::runtime(shape_template),
+                                       contour_file.getPath(),
+                                       "void",
+                                       "disk",
+                                       "hole"));
 
     // Create an initial background material set to 1 everywhere
     std::map<std::string, mfem::GridFunction*> initialGridFunctions;
@@ -863,8 +805,8 @@ shapes:
       initialGridFunctions["void"] = vf;
     }
 
-    this->validateShapeFile(shape_file.getFileName());
-    this->initializeShaping(shape_file.getFileName(), initialGridFunctions);
+    this->validateShapeFile(shape_file.getPath());
+    this->initializeShaping(shape_file.getPath(), initialGridFunctions);
     this->runShaping();
 
     // check that the result has a volume fraction field associated with the circle material
@@ -886,12 +828,12 @@ shapes:
 
   // Set background and inner hole materials to 'void'
   {
-    ScopedTemporaryFile shape_file(axom::fmt::format("{}.yaml", testname),
-                                   axom::fmt::format(axom::fmt::runtime(shape_template),
-                                                     contour_file.getFileName(),
-                                                     "void",
-                                                     "disk",
-                                                     "void"));
+    fs::TempFile shape_file(testname, ".yaml");
+    shape_file.write(axom::fmt::format(axom::fmt::runtime(shape_template),
+                                       contour_file.getPath(),
+                                       "void",
+                                       "disk",
+                                       "void"));
 
     // Create an initial background material set to 1 everywhere
     std::map<std::string, mfem::GridFunction*> initialGridFunctions;
@@ -902,8 +844,8 @@ shapes:
       initialGridFunctions["void"] = vf;
     }
 
-    this->validateShapeFile(shape_file.getFileName());
-    this->initializeShaping(shape_file.getFileName(), initialGridFunctions);
+    this->validateShapeFile(shape_file.getPath());
+    this->initializeShaping(shape_file.getPath(), initialGridFunctions);
     this->runShaping();
 
     // check that the result has a volume fraction field associated with the circle material
@@ -949,9 +891,8 @@ shapes:
   does_not_replace: [{1}]
 )";
 
-  ScopedTemporaryFile shape_file(
-    axom::fmt::format("{}.yaml", testname),
-    axom::fmt::format(axom::fmt::runtime(shape_template), "void", "left", "odds"));
+  fs::TempFile shape_file(testname, ".yaml");
+  shape_file.write(axom::fmt::format(axom::fmt::runtime(shape_template), "void", "left", "odds"));
 
   if(very_verbose_output)
   {
@@ -979,8 +920,8 @@ shapes:
     initialGridFunctions["odds"] = vf;
   }
 
-  this->validateShapeFile(shape_file.getFileName());
-  this->initializeShaping(shape_file.getFileName(), initialGridFunctions);
+  this->validateShapeFile(shape_file.getPath());
+  this->initializeShaping(shape_file.getPath(), initialGridFunctions);
   this->runShaping();
 
   // check that the result has a volume fraction field associated with the circle material
@@ -1045,11 +986,11 @@ shapes:
   replaces: [void, hole]
 )";
 
-  ScopedTemporaryFile contour_file(axom::fmt::format("{}.contour", testname), unit_circle_contour);
+  fs::TempFile contour_file(testname, ".contour");
+  contour_file.write(unit_circle_contour);
 
-  ScopedTemporaryFile shape_file(
-    axom::fmt::format("{}.yaml", testname),
-    axom::fmt::format(axom::fmt::runtime(shape_template), contour_file.getFileName()));
+  fs::TempFile shape_file(testname, ".yaml");
+  shape_file.write(axom::fmt::format(axom::fmt::runtime(shape_template), contour_file.getPath()));
 
   if(very_verbose_output)
   {
@@ -1087,8 +1028,8 @@ shapes:
   // The interior hole is within radius .5, but is replaced by left and odds
   // The odds material is all cells w/ odd index, but not covering 'left' or 'disk'
   // The end result for the void background is everything that's left
-  this->validateShapeFile(shape_file.getFileName());
-  this->initializeShaping(shape_file.getFileName(), initialGridFunctions);
+  this->validateShapeFile(shape_file.getPath());
+  this->initializeShaping(shape_file.getPath(), initialGridFunctions);
   this->runShaping();
 
   // check that the result has the correct volume fractions
@@ -1160,19 +1101,19 @@ shapes:
   const std::string double_underscored_shape_name {"double_underscored_shape"};
   const std::string double_underscored_mat_name {"double_underscored_mat"};
 
-  ScopedTemporaryFile contour_file(axom::fmt::format("{}.contour", testname),
-                                   unit_semicircle_contour);
+  fs::TempFile contour_file(testname, ".contour");
+  contour_file.write(unit_semicircle_contour);
 
-  ScopedTemporaryFile shape_file(axom::fmt::format("{}.yaml", testname),
-                                 axom::fmt::format(axom::fmt::runtime(shape_template),
-                                                   contour_file.getFileName(),
-                                                   radius,
-                                                   shape_name,
-                                                   mat_name,
-                                                   underscored_shape_name,
-                                                   underscored_mat_name,
-                                                   double_underscored_shape_name,
-                                                   double_underscored_mat_name));
+  fs::TempFile shape_file(testname, ".yaml");
+  shape_file.write(axom::fmt::format(axom::fmt::runtime(shape_template),
+                                     contour_file.getPath(),
+                                     radius,
+                                     shape_name,
+                                     mat_name,
+                                     underscored_shape_name,
+                                     underscored_mat_name,
+                                     double_underscored_shape_name,
+                                     double_underscored_mat_name));
 
   if(very_verbose_output)
   {
@@ -1180,8 +1121,8 @@ shapes:
     SLIC_INFO("Shape file: \n" << shape_file.getFileContents());
   }
 
-  this->validateShapeFile(shape_file.getFileName());
-  this->initializeShaping(shape_file.getFileName());
+  this->validateShapeFile(shape_file.getPath());
+  this->initializeShaping(shape_file.getPath());
 
   this->runShaping();
 
@@ -1246,16 +1187,17 @@ shapes:
   const std::string sphere_material = "vaccum";
   const std::string sphere_path = axom::fmt::format("{}/quest/unit_sphere.stl", AXOM_DATA_DIR);
 
-  ScopedTemporaryFile contour_file(axom::fmt::format("{}.contour", testname), unit_circle_contour);
+  fs::TempFile contour_file(testname, ".contour");
+  contour_file.write(unit_circle_contour);
 
-  ScopedTemporaryFile shape_file(axom::fmt::format("{}.yaml", testname),
-                                 axom::fmt::format(axom::fmt::runtime(shape_template),
-                                                   contour_file.getFileName(),
-                                                   sphere_path,
-                                                   radius,
-                                                   background_material,
-                                                   circle_material,
-                                                   sphere_material));
+  fs::TempFile shape_file(testname, ".yaml");
+  shape_file.write(axom::fmt::format(axom::fmt::runtime(shape_template),
+                                     contour_file.getPath(),
+                                     sphere_path,
+                                     radius,
+                                     background_material,
+                                     circle_material,
+                                     sphere_material));
 
   if(very_verbose_output)
   {
@@ -1271,8 +1213,8 @@ shapes:
     initialGridFunctions[background_material] = vf;
   }
 
-  this->validateShapeFile(shape_file.getFileName());
-  this->initializeShaping(shape_file.getFileName(), initialGridFunctions);
+  this->validateShapeFile(shape_file.getPath());
+  this->initializeShaping(shape_file.getPath(), initialGridFunctions);
 
   // set projector from 2D mesh points to 3D query points within STL
   this->m_shaper->setPointProjector23([](Point2D pt) { return Point3D {pt[0], pt[1], 0.}; });
@@ -1319,9 +1261,8 @@ shapes:
   const std::string tet_material = "steel";
   const std::string tet_path = axom::fmt::format("{}/quest/tetrahedron.stl", AXOM_DATA_DIR);
 
-  ScopedTemporaryFile shape_file(
-    axom::fmt::format("{}.yaml", testname),
-    axom::fmt::format(axom::fmt::runtime(shape_template), tet_material, tet_path));
+  fs::TempFile shape_file(testname, ".yaml");
+  shape_file.write(axom::fmt::format(axom::fmt::runtime(shape_template), tet_material, tet_path));
 
   if(very_verbose_output)
   {
@@ -1329,8 +1270,8 @@ shapes:
     SLIC_INFO("Shape file: \n" << shape_file.getFileContents());
   }
 
-  this->validateShapeFile(shape_file.getFileName());
-  this->initializeShaping(shape_file.getFileName());
+  this->validateShapeFile(shape_file.getPath());
+  this->initializeShaping(shape_file.getPath());
   this->runShaping();
 
   // Check that the result has a volume fraction field associated with the tetrahedron material
@@ -1396,16 +1337,15 @@ shapes:
   const std::string tet_material = "steel";
   const std::string tet_path = axom::fmt::format("{}/quest/tetrahedron.stl", AXOM_DATA_DIR);
 
-  ScopedTemporaryFile shape_file(
-    axom::fmt::format("{}.yaml", testname),
-    axom::fmt::format(axom::fmt::runtime(shape_template), tet_material, tet_path));
+  fs::TempFile shape_file(testname, ".yaml");
+  shape_file.write(axom::fmt::format(axom::fmt::runtime(shape_template), tet_material, tet_path));
 
   if(very_verbose_output)
   {
     SLIC_INFO("Shape file: \n" << shape_file.getFileContents());
   }
 
-  this->validateShapeFile(shape_file.getFileName());
+  this->validateShapeFile(shape_file.getPath());
 
   // Create initial background materials based on octant attributes
   // Octants were offset by one since mfem doesn't allow setting attribute to zero
@@ -1421,7 +1361,7 @@ shapes:
     }
   }
 
-  this->initializeShaping(shape_file.getFileName(), initialGridFunctions);
+  this->initializeShaping(shape_file.getPath(), initialGridFunctions);
   this->runShaping();
 
   // Check that the result has a volume fraction field associated with the tetrahedron material
@@ -1510,15 +1450,15 @@ shapes:
 
   const std::string tet_path = axom::fmt::format("{}/quest/tetrahedron.stl", AXOM_DATA_DIR);
 
-  ScopedTemporaryFile shape_file(axom::fmt::format("{}.yaml", testname),
-                                 axom::fmt::format(axom::fmt::runtime(shape_template), tet_path));
+  fs::TempFile shape_file(testname, ".yaml");
+  shape_file.write(axom::fmt::format(axom::fmt::runtime(shape_template), tet_path));
 
   if(very_verbose_output)
   {
     SLIC_INFO("Shape file: \n" << shape_file.getFileContents());
   }
 
-  this->validateShapeFile(shape_file.getFileName());
+  this->validateShapeFile(shape_file.getPath());
 
   // Create initial background materials based on octant attributes
   std::map<std::string, mfem::GridFunction*> initialGridFunctions;
@@ -1533,7 +1473,7 @@ shapes:
     }
   }
 
-  this->initializeShaping(shape_file.getFileName(), initialGridFunctions);
+  this->initializeShaping(shape_file.getPath(), initialGridFunctions);
   this->runShaping();
 
   // Check that the result has a volume fraction field associated with the tetrahedron material
@@ -1585,17 +1525,16 @@ shapes:
   const std::string tet_material = "steel";
   const std::string tet_path = axom::fmt::format("{}/quest/tetrahedron.stl", AXOM_DATA_DIR);
 
-  ScopedTemporaryFile shape_file(
-    axom::fmt::format("{}.yaml", testname),
-    axom::fmt::format(axom::fmt::runtime(shape_template), tet_material, tet_path));
+  fs::TempFile shape_file(testname, ".yaml");
+  shape_file.write(axom::fmt::format(axom::fmt::runtime(shape_template), tet_material, tet_path));
 
   if(very_verbose_output)
   {
     SLIC_INFO("Shape file: \n" << shape_file.getFileContents());
   }
 
-  this->validateShapeFile(shape_file.getFileName());
-  this->initializeShaping(shape_file.getFileName());
+  this->validateShapeFile(shape_file.getPath());
+  this->initializeShaping(shape_file.getPath());
 
   // check that we can set several projectors in 2D and 3D
   // uses simplest projectors, e.g. identity in 2D and 3D
@@ -1642,17 +1581,16 @@ shapes:
   const std::string tet_material = "steel";
   const std::string tet_path = axom::fmt::format("{}/quest/tetrahedron.stl", AXOM_DATA_DIR);
 
-  ScopedTemporaryFile shape_file(
-    axom::fmt::format("{}.yaml", testname),
-    axom::fmt::format(axom::fmt::runtime(shape_template), tet_material, tet_path));
+  fs::TempFile shape_file(testname, ".yaml");
+  shape_file.write(axom::fmt::format(axom::fmt::runtime(shape_template), tet_material, tet_path));
 
   if(very_verbose_output)
   {
     SLIC_INFO("Shape file: \n" << shape_file.getFileContents());
   }
 
-  this->validateShapeFile(shape_file.getFileName());
-  this->initializeShaping(shape_file.getFileName());
+  this->validateShapeFile(shape_file.getPath());
+  this->initializeShaping(shape_file.getPath());
 
   // scale input points by a factor of 1/2 in each dimension
   this->m_shaper->setPointProjector33([](const Point3D& pt) {
@@ -1701,14 +1639,14 @@ shapes:
 
   const std::string circle_material = "steel";
 
-  ScopedTemporaryFile contour_file(axom::fmt::format("{}.contour", testname),
-                                   unit_semicircle_contour);
+  fs::TempFile contour_file(testname, ".contour");
+  contour_file.write(unit_semicircle_contour);
 
-  ScopedTemporaryFile shape_file(axom::fmt::format("{}.yaml", testname),
-                                 axom::fmt::format(axom::fmt::runtime(shape_template),
-                                                   circle_material,
-                                                   contour_file.getFileName(),
-                                                   radius));
+  fs::TempFile shape_file(testname, ".yaml");
+  shape_file.write(axom::fmt::format(axom::fmt::runtime(shape_template),
+                                     circle_material,
+                                     contour_file.getPath(),
+                                     radius));
 
   if(very_verbose_output)
   {
@@ -1716,8 +1654,8 @@ shapes:
     SLIC_INFO("Shape file: \n" << shape_file.getFileContents());
   }
 
-  this->validateShapeFile(shape_file.getFileName());
-  this->initializeShaping(shape_file.getFileName());
+  this->validateShapeFile(shape_file.getPath());
+  this->initializeShaping(shape_file.getPath());
 
   // set projector from 3D points to axisymmetric plane
   this->m_shaper->setPointProjector32([](Point3D pt) {
@@ -1778,16 +1716,16 @@ shapes:
   const std::string sphere_material = "void";
   const std::string sphere_path = axom::fmt::format("{}/quest/unit_sphere.stl", AXOM_DATA_DIR);
 
-  ScopedTemporaryFile contour_file(axom::fmt::format("{}.contour", testname),
-                                   unit_semicircle_contour);
+  fs::TempFile contour_file(testname, ".contour");
+  contour_file.write(unit_semicircle_contour);
 
-  ScopedTemporaryFile shape_file(axom::fmt::format("{}.yaml", testname),
-                                 axom::fmt::format(axom::fmt::runtime(shape_template),
-                                                   contour_file.getFileName(),
-                                                   sphere_path,
-                                                   radius,
-                                                   circle_material,
-                                                   sphere_material));
+  fs::TempFile shape_file(testname, ".yaml");
+  shape_file.write(axom::fmt::format(axom::fmt::runtime(shape_template),
+                                     contour_file.getPath(),
+                                     sphere_path,
+                                     radius,
+                                     circle_material,
+                                     sphere_material));
 
   if(very_verbose_output)
   {
@@ -1795,8 +1733,8 @@ shapes:
     SLIC_INFO("Shape file: \n" << shape_file.getFileContents());
   }
 
-  this->validateShapeFile(shape_file.getFileName());
-  this->initializeShaping(shape_file.getFileName());
+  this->validateShapeFile(shape_file.getPath());
+  this->initializeShaping(shape_file.getPath());
 
   // set projector from 3D points to axisymmetric plane
   this->m_shaper->setPointProjector32([](Point3D pt) {
@@ -1865,11 +1803,12 @@ shapes:
   // clang-format on
 
   const std::string tet_material = "steel";
-  ScopedTemporaryFile tet_file(axom::fmt::format("{}.proe", testname), proe_tet_str);
+  fs::TempFile tet_file(testname, ".proe");
+  tet_file.write(proe_tet_str);
 
-  ScopedTemporaryFile shape_file(
-    axom::fmt::format("{}.yaml", testname),
-    axom::fmt::format(axom::fmt::runtime(shape_template), tet_file.getFileName(), tet_material));
+  fs::TempFile shape_file(testname, ".yaml");
+  shape_file.write(
+    axom::fmt::format(axom::fmt::runtime(shape_template), tet_file.getPath(), tet_material));
 
   if(very_verbose_output)
   {
@@ -1877,14 +1816,14 @@ shapes:
     SLIC_INFO("Shape file: \n" << shape_file.getFileContents());
   }
 
-  this->validateShapeFile(shape_file.getFileName());
+  this->validateShapeFile(shape_file.getPath());
 
   // exercise the point projector by running at several XY planes
   for(const double z : {-1.2, -.9, -.75, -.1, 0., .25, 1. / 3., .5})
   {
     this->resetShaping();
 
-    this->initializeShaping(shape_file.getFileName());
+    this->initializeShaping(shape_file.getPath());
 
     primal::Plane<double, 3> plane({0, 0, 1}, z);
     const auto polygon = slice(tet, plane);
@@ -1974,13 +1913,14 @@ piece = line(end=start)
                         bb.getMax()[0],
                         bb.getMax()[1]);
 
-    ScopedTemporaryFile contour_file(axom::fmt::format("{}.contour", testname), contour_str);
+    fs::TempFile contour_file(testname, ".contour");
+    contour_file.write(contour_str);
 
-    ScopedTemporaryFile shape_file(axom::fmt::format("{}.yaml", testname),
-                                   axom::fmt::format(axom::fmt::runtime(shape_template),
-                                                     rect_shape,
-                                                     rect_material,
-                                                     contour_file.getFileName()));
+    fs::TempFile shape_file(testname, ".yaml");
+    shape_file.write(axom::fmt::format(axom::fmt::runtime(shape_template),
+                                       rect_shape,
+                                       rect_material,
+                                       contour_file.getPath()));
 
     if(very_verbose_output)
     {
@@ -1988,12 +1928,12 @@ piece = line(end=start)
       SLIC_INFO("Shape file: \n" << shape_file.getFileContents());
     }
 
-    this->validateShapeFile(shape_file.getFileName());
+    this->validateShapeFile(shape_file.getPath());
 
     for(int qorder : {3, 4, 5, 6, 7, 8, 9})
     {
       this->resetShaping();
-      this->initializeShaping(shape_file.getFileName());
+      this->initializeShaping(shape_file.getPath());
 
       this->m_shaper->setVolumeFractionOrder(0);
       this->m_shaper->setQuadratureOrder(qorder);